Rewrote the main algorithm functions
To stop working with the segments ranges defined in
comet/sma/segments.d.
This is because of the new phylogenetic tree functionality. The passed
phylogenetic tree will hold references to the sequences extracted from
the fasta files directly. Now, while building the smtree so that its
structure is a duplication speciation tree, the program directly maps
the leaves to the corresponding sequences. This results in a good
speedup when setting the smtree leaves on every iteration.